+++ app/ide-desktop/lib/dashboard/src/text/english.json
@@ -578,6 +578,7 @@
   "tryAgain": "Try again",
   "arbitraryErrorTitle": "An error occurred",
   "arbitraryErrorSubtitle": "Please try again or contact the administrators.",
+  "arbitraryFormErrorMessage": "Something went wrong while submitting the form. Please try again or contact the administrators.",
 
   "subscribeSubmit": "Subscribe",
   "BankCardLabel": "Bank Card",
